requirements.txt
fastapi
uvicorn
motor
pydantic[email]
python-dotenv
cloudinary
mangum
bcrypt
pymongo
PyJWT